+++ src/commonMain/kotlin/uk/gov/hmrc/calculator/model/pension/PensionAllowances.kt
@@ -30,20 +30,17 @@ internal object PensionAllowances {
         TaxYear.TWENTY_TWENTY_FOUR -> pensionAllowance2024()
     }
 
-    private fun pensionAllowance2020() = PensionAllowance(1073100.0, 40000.0)
+    private fun pensionAllowance2020() = PensionAllowance(40000.0)
 
-    private fun pensionAllowance2021() = PensionAllowance(1073100.0, 40000.0)
+    private fun pensionAllowance2021() = PensionAllowance(40000.0)
 
-    private fun pensionAllowance2022() = PensionAllowance(1073100.0, 40000.0)
+    private fun pensionAllowance2022() = PensionAllowance(40000.0)
 
-    private fun pensionAllowance2023() = PensionAllowance(1073100.0, 60000.0)
+    private fun pensionAllowance2023() = PensionAllowance(60000.0)
 
-    private fun pensionAllowance2023JanuaryRevised() = PensionAllowance(1073100.0, 60000.0)
+    private fun pensionAllowance2023JanuaryRevised() = PensionAllowance(60000.0)
 
-    private fun pensionAllowance2024() = PensionAllowance(1073100.0, 60000.0)
+    private fun pensionAllowance2024() = PensionAllowance(60000.0)
 
-    internal data class PensionAllowance(
-        val standardLifetimeAllowance: Double,
-        val annualAllowance: Double
-    )
+    internal data class PensionAllowance(val annualAllowance: Double)
 }